+++ models/menu.py
@@ -7,7 +7,7 @@ class Menu(models.Model):
     name = fields.Char(string="Name", required=True)
     dish_ids = fields.Many2many('restaurantmng.dish', string="Dishes", required=True)
     drink_id = fields.Many2one('restaurantmng.drink', string="Drink", required=True)
-    price = fields.Integer(compute='_calculate_price', string="Price (in €)")
+    price = fields.Float(compute='_calculate_price', string="Price (in €)")
 
     @api.depends('dish_ids', 'drink_id')
     def _calculate_price(self):
